Add a `get_defaulting` helper and break out utils
